Event saving: fix problem with storing nested events
Having two sub-groups OR'ed, each consisting of two operators AND'ed,
but with no operator in the root group, yielded the following nesting:

   ' or ((criteria[0] and criteria[1])) or ((criteria[2] and criteria[3])))'

which condensed to

   '((criteria[0] and criteria[1])) or ((criteria[2] and criteria[3])))'

Which failed to execute in the backend.

This patch ensures we have an additional ( instead of  ' or ', if there
is no operator in the root group.